+++ src/Validation/RulesGatherer.php
@@ -2,7 +2,6 @@
 
 namespace Nuwave\Lighthouse\Validation;
 
-use Closure;
 use Illuminate\Contracts\Validation\Rule;
 use Illuminate\Support\Collection;
 use Illuminate\Validation\ValidationRuleParser;
@@ -216,17 +215,13 @@ class RulesGatherer
      *
      * @param  array<int, mixed>  $rules
      * @param  array<int|string>  $argumentPath
-     * @return array<int, array<int, mixed>|\Illuminate\Contracts\Validation\Rule|Closure>
+     * @return array<int, array<int, mixed>|object>
      */
     protected function qualifyArgumentReferences(array $rules, array $argumentPath): array
     {
         return array_map(
             static function ($rule) use ($argumentPath) {
-                if ($rule instanceof Rule) {
-                    return $rule;
-                }
-
-                if ($rule instanceof Closure) {
+                if (is_object($rule)) {
                     return $rule;
                 }